requirements.txt

# Core web
flask==3.0.3
gunicorn==22.0.0

# YOLOv8 + deps
ultralytics==8.3.26
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy>=1.23.0
